pyproject.toml
[project]
name = "pressing-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
